templates/bot-template/requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
